# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Image Recognition Microservices App
  description: The app allows users to run inference on images for image detection and provides additional functionalities like video streaming, user management, and API usage tracking.
  version: v1
servers:
  - url: https://isa-singh.azurewebsites.net/
paths:
  /login:
    get:
      summary: Display the login page
      description: Returns the login page HTML.
      responses:
        '200':
          description: Successfully retrieved the login page.
    post:
      summary: Authenticate user
      description: Authenticates the user with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully authenticated.
        '401':
          description: Unauthorized.
  /register:
    post:
      summary: Register a new user
      description: Allows a user to register with email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Successfully registered.
        '401':
          description: Unauthorized.
    get:
      summary: Display the registration page
      description: Returns the registration page HTML.
      responses:
        '200':
          description: Successfully retrieved the registration page.
...
  /reset-password:
    post:
      summary: Reset password
      description: Resets the password using a JWT and new password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                token:
                  type: string
      responses:
        '200':
          description: Successfully reset password.
        '401':
          description: Unauthorized.
...
  /query:
    post:
      summary: Query data
      description: Endpoint for querying data from the database (details TBD).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true # Modify with exact query structure when known
      responses:
        '200':
          description: Successfully queried data.
        '401':
          description: Unauthorized.
  /delete:
    delete:
      summary: Delete a user
      description: Deletes a user from the database based on the provided email.
      security:
        - bearerAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
          description: Email of the user to delete.
      responses:
        '200':
          description: Successfully deleted user.
        '404':
          description: User not found.
        '401':
          description: Unauthorized.
  /update-usage:
    put:
      summary: Update user API usage
      description: Updates a user's API usage counter.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                counter:
                  type: integer
                  description: New usage value.
      responses:
        '200':
          description: Successfully updated usage.
        '404':
          description: User not found.
        '401':
          description: Unauthorized.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT